SamlResponse signature validated
The ServiceProviderSigningCertificate was used for both decryption (line _xmlProvider.GetAssertion(element, signingCertificate.PrivateKey);) and signature validation (line var assertion = new Saml2Assertion(assertionElement ...)

The SamlResponse is signed with the IdP certificate, not the SP certificate. (tested on an ADFS 2.0 server)

identityProviderName parameter is added to ISamlValidator in order to retrieve the IdP signing Certificate.
The existing variable signingCertificate is renamed encryptionCertificate because it is used to decrypt the assertion (if the assertion is encrypted)